monitoring/requirements.txt
aiohttp==3.7.3
aiopg==1.1.0
async-timeout==3.0.1
asyncstdlib==3.9.0
attrs==20.3.0
chardet==3.0.4
idna==3.1
iniconfig==1.1.1
multidict==5.1.0
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pydantic==1.7.3
pyparsing==2.4.7
pytest==6.2.2
pytest-asyncio==0.14.0
PyYAML==5.3.1
toml==0.10.2
typing-extensions==3.7.4.3
yarl==1.6.3
